Rename Mongo tags providers to align with existing tags providers.
Fixes gh-2541
